Add CleanCLIP and TIJO defense baselines.
CleanCLIP (Bansal et al. 2023, NeurIPS):
- sharpguard/baselines/cleanclip.py
- Adapts the inmodal augmentation contrastive loss (their key contribution)
  to OpenVLA: image embed (from vision tower) ↔ augmented image embed
  via InfoNCE, weight 0.5. Crossmodal omitted because OpenVLA fuses image
  + text early (per their architecture).

TIJO (Sur et al. 2023):
- sharpguard/baselines/tijo.py
- Trigger inversion via gradient optimization: optimize a small image
  perturbation (L∞ ≤ 0.3) so model(clean_image + x) emits the malicious
  action token sequence. Detect training samples by similarity to the
  inverted trigger.

Headline now compares 7 defense rows (FT, FT-SAM, FT-AC, Fine-pruning,
FT-Attn, CleanCLIP, TIJO) against SharpGuard mech-A and mech-B.

Also fixes:
- attention-entropy detector: cast tensor to float32 before torch.quantile
  (bf16 not supported by quantile in PyTorch 2.4).

+++ sharpguard/baselines/attention_entropy.py
@@ -109,8 +109,7 @@ def detect_poison_attention(
             ent = -(text_to_vis * (text_to_vis.clamp_min(1e-12).log())).sum(dim=-1)  # [B, T_text]
             entropies.append(ent.mean(dim=-1).cpu())
 
-    ent_per_sample = torch.cat(entropies, dim=0)              # [N]
-
+    ent_per_sample = torch.cat(entropies, dim=0).float()      # [N], force fp32 for quantile
     # Replace any NaNs with median so they don't get spuriously flagged.
     nan_mask = torch.isnan(ent_per_sample)
     if nan_mask.any():
